exceptions added

+++ src/products/products.service.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Injectable, NotFoundException } from "@nestjs/common";
+import { BadRequestException, Injectable, NotFoundException, UnauthorizedException } from "@nestjs/common";
 import { Product } from "./products.entity";
 import { InjectRepository } from "@nestjs/typeorm";
 import { CreateProductDto } from "./dto/create-product.dto";
@@ -8,16 +8,15 @@ import { UsersService } from "../users/users.service";
 @Injectable()
 export class ProductsService {
 
-  constructor(@InjectRepository(Product) private productRepository: Repository<Product>, private readonly usersService: UsersService,) {
+  constructor(@InjectRepository(Product) private productRepository: Repository<Product>, private readonly usersService: UsersService) {
 
   }
 
   async createProduct(dto: CreateProductDto, userId: number): Promise<Product> {
     const userExists = await this.usersService.userExists(userId);
     if (!userExists) {
-      throw new BadRequestException('User not verified');
+      throw new UnauthorizedException("User not verified");
     }
-
     const product = this.productRepository.create({ ...dto, user: userId });
     await this.productRepository.save(product);
     return product;
@@ -27,15 +26,20 @@ export class ProductsService {
     return this.productRepository.find();
   }
 
-  getProductById(id: number): Promise<Product> {
-    return this.productRepository.findOneBy({ id });
+  async getProductById(id: number): Promise<Product> {
+    const product = await this.productRepository.findOneBy({ id });
+    if (!product) {
+      throw new NotFoundException(`Product with ID ${id} not found`);
+    }
+    return product;
   }
 
   async deleteProductById(id: number): Promise<void> {
-    try{
-      await this.productRepository.delete(id);
-    } catch (error){
+    const product = await this.productRepository.findOneBy({ id });
+    if (!product) {
       throw new NotFoundException(`Product with ID ${id} not found`);
+    } else {
+      await this.productRepository.delete(id);
     }
   }
